+++ src/allowMessagesFromGroup.ts
@@ -1,6 +1,6 @@
 import bridge from '@vkontakte/vk-bridge';
 
-/*
+/**
  * Wrapper to VKWebAppAllowMessagesFromGroup. Asks to allow messages from group
  * @param {number} group_id ID of VK group. Taken from Window by default
  * @returns {boolean} Returns true if success