+++ src/index.ts
@@ -16,6 +16,7 @@ import {
 import vkApi from './vkApi';
 import initializeVkApp from './initializeVkApp';
 import checkIOS from './checkIOS';
+import checkMobile from './checkMobile';
 
 export {
   addToFavorites,
@@ -33,6 +34,7 @@ export {
   vibrateAsSelection,
   vkApi,
   initializeVkApp,
+  checkMobile,
   checkIOS,
 };